Moved Admin API schema sources from lib to src
ref https://github.com/TryGhost/Ghost/pull/29449

Separated the directory relocation from the TypeScript conversion so Git records the migrated files at full similarity and retains their ancestry.
