+++ packages/admin-api-schema/src/schemas/images-upload.json
