+++ packages/admin-api-schema/src/schemas/images.json
