+++ packages/admin-api-schema/src/schemas/labels-add.json
