+++ packages/admin-api-schema/src/schemas/labels-edit.json
