+++ packages/admin-api-schema/src/schemas/labels.json
