+++ packages/admin-api-schema/src/schemas/media-upload.json
