+++ packages/admin-api-schema/src/schemas/media.json
