+++ packages/admin-api-schema/src/schemas/posts.json
