+++ packages/admin-api-schema/src/schemas/products-add.json
