+++ packages/admin-api-schema/src/schemas/products-edit.json
