+++ packages/admin-api-schema/src/schemas/products.json
