+++ packages/admin-api-schema/src/schemas/snippets-add.json
