+++ packages/admin-api-schema/src/schemas/snippets-edit.json
