+++ packages/admin-api-schema/src/schemas/snippets.json
