+++ packages/admin-api-schema/src/schemas/tags-add.json
