+++ packages/admin-api-schema/src/schemas/tags-edit.json
