+++ packages/admin-api-schema/src/schemas/tags.json
