+++ packages/admin-api-schema/src/schemas/webhooks-edit.json
